+++ lsp/test/text_document_tests.ml
@@ -1,6 +1,7 @@
 open Lsp
 open Lsp.Types
 module List = ListLabels
+module String = StringLabels
 
 let tuple_range start end_ =
   { Range.start =
@@ -30,8 +31,8 @@ let test_general text changes =
     in
     Text_document.text td
   in
-  let utf16 = test `UTF16 in
   let utf8 = test `UTF8 in
+  let utf16 = test `UTF16 in
   let printf = Printf.printf in
   if String.equal utf16 utf8 then printf "result: %s\n" (String.escaped utf8)
   else (
@@ -62,9 +63,51 @@ let%expect_test "no range" =
   [%expect {|
     result: XXXX |}]
 
+let%expect_test "char by char" =
+  test_multiple
+    ""
+    [ (tuple_range (0, 0) (0, 0), "f")
+    ; (tuple_range (0, 1) (0, 1), "o")
+    ; (tuple_range (0, 2) (0, 2), "o")
+    ];
+  [%expect {|
+    result: foo |}]
+
+let%expect_test "char by char - 2" =
+  test_multiple
+    "char by char - 2\n"
+    [ (tuple_range (1, 10) (1, 10), "b")
+    ; (tuple_range (1, 10) (1, 10), "a")
+    ; (tuple_range (1, 10) (1, 10), "r")
+    ; (tuple_range (1, 1) (1, 2), "")
+    ];
+  [%expect {|
+    result: char by char - 2\nbr |}]
+
+let%expect_test "char by char - 3" =
+  test_multiple
+    "first line skip\nchar by char - 2\n"
+    [ (tuple_range (1, 4) (1, 5), "")
+    ; (tuple_range (1, 3) (1, 4), "")
+    ; (tuple_range (1, 3) (1, 3), "x")
+    ];
+  [%expect {|
+    result: first line skip\nchaxby char - 2\n |}]
+
+let%expect_test "insert last" =
+  test "x" (tuple_range (0, 1) (0, 1)) ~change:"y";
+  [%expect {|
+    result: xy |}];
+  test "x\ny" (tuple_range (1, 1) (1, 1)) ~change:"z";
+  [%expect {|
+    result: x\nyz |}];
+  test "x\ny" (tuple_range (1, 10) (1, 10)) ~change:"z";
+  [%expect {|
+    result: x\nyz |}]
+
 let%expect_test "replace second line" =
   let range = tuple_range (1, 0) (2, 0) in
-  test "foo\n\bar\nbaz\n" range ~change:"XXXX\n";
+  test "foo\nbar\nbaz\n" range ~change:"XXXX\n";
   [%expect {|
     result: foo\nXXXX\nbaz\n |}]
 
@@ -75,10 +118,14 @@ let%expect_test "edit in second line" =
     result: foo\nb-XXX-r\nbaz\n |}]
 
 let%expect_test "insert at the end" =
-  let range = tuple_range (3, 1) (4, 0) in
-  test "foo\n\bar\nbaz\n" range ~change:"XXX";
+  let range = tuple_range (3, 0) (3, 0) in
+  test "foo\nbar\nbaz\n" range ~change:"XXX";
   [%expect {|
-   result: foo\n\bar\nbaz\nXXX |}]
+    result: foo\nbar\nbaz\nXXX |}];
+  let range = tuple_range (3, 0) (4, 0) in
+  test "foo\nbar\nbaz\n" range ~change:"XXX";
+  [%expect {|
+    result: foo\nbar\nbaz\nXXX |}]
 
 let%expect_test "insert at the beginning" =
   let range = tuple_range (0, 0) (0, 0) in
@@ -131,7 +178,8 @@ let%expect_test "remove text" =
 
 let%expect_test "remove newline - 1" =
   test "\n" (tuple_range (0, 0) (0, 1)) ~change:"";
-  [%expect {| result: \n |}]
+  [%expect {|
+    result: \n |}]
 
 let%expect_test "remove newlines - 2" =
   test_multiple "\nXXX\n" [ (tuple_range (0, 0) (0, 1), "") ];
@@ -142,8 +190,17 @@ let%expect_test "remove newlines - 3" =
   test_multiple
     "\nXXX\n\n"
     [ (tuple_range (0, 0) (0, 1), ""); (tuple_range (0, 1) (0, 2), "") ];
-  [%expect
-    {|
-    [FAILURE] utf16 and utf8 disagree
-    utf16: XX\n\n
-    utf8:  \nXXX\n\n |}]
+  [%expect {|
+    result: \nXXX\n\n |}]
+
+let%expect_test "update when inserting a line at the end of the doc" =
+  test "let x = 1;\n\nlet y = 2;" (tuple_range (2, 10) (2, 10)) ~change:"\n-ZZZ";
+  [%expect {|
+    result: let x = 1;\n\nlet y = 2;\n-ZZZ |}]
+
+let%expect_test "update when inserting a line at the end of the doc" =
+  test_multiple
+    "1\n2\n3\n"
+    [ (tuple_range (1, 9) (1, 9), "l"); (tuple_range (1, 9) (1, 10), "") ];
+  [%expect {|
+    result: 1\n2l\n3\n |}]