+++ CHANGES.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+## Fixes
+
+- Fix a document syncing issue when utf-16 is the position encoding (#1004)
+
 # 1.15.1
 
 ## Fixes
@@ -6,8 +12,6 @@
   running ([#1005](https://github.com/ocaml/ocaml-lsp/pull/1005), fixes
   [#941](https://github.com/ocaml/ocaml-lsp/issues/941),
   [#1003](https://github.com/ocaml/ocaml-lsp/issues/1003))
-
-# 1.15.0
 
 ## Features